Separated initialization from ST_grid and added header files (Large commit)
(issue #342)
In this commit I added three files: ST_grid.h, ST_initialization.c, and ST_initialization.h. See below for a description of each:

ST_initialization.c:
* Contains all of the functions from ST_grid.c that were related to initialization. I did this to separate the two features.
* ST_grid.c now contains ONLY the functions that deal with gridded mode and the progress bar.

ST_initialization.h:
* Contains declarations for the functions we want to be called from other files.
* Contains the Status enumerator (previously in ST_grid.c) so all files that include ST_initialization.h can use it.
* Contains the declarations for initializationCells, initializationMethod, and DuringInitialization.
* initializationMethod and DuringInitialization were removed from ST_globals.h. This way it is clear that the variables belong to the initialization module.

ST_grid.h:
* Moved the declarations for a few functions here. These functions could all be reasonably called from other files. This allowed me to remove the declaration of runGrid() from ST_main.c and the declarations of load_cell and unload_cell from ST_initialization.c.
* Removed all functions related to initialization.
* Moved the definitions of all structs and enumerators from ST_grid.c to this file.

I modified 6 other files:

makefile:
* Added rule for ST_initialization.o.
* Removed all header files from the rule for ST_grid.o. Header files are NOT needed in makefile rules.

ST_globals.h:
* Removed UseSeedDispersal, InitializationMethod and DuringInitialization.
* These variables are now defined in the grid and initialization headers.

ST_grid.c:
* Moved module-level function declarations to the header file.
* Moved all structs and enumerators to the header files.
* Removed time.h from include statements. It is not needed.
* Moved some of the include statements to ST_grid.h. The include statements that I moved are all required to use the module.

ST_main.c:
* Added and include statement for ST_initialization.h. ST_initialization.h contains ST_grid.h so we get both of them.
* Removed redeclaration of runGrid().
* Removed declaration of UseSeedDispersal, InitializationMethod, and DuringInitialization. All of these are now declared in the header files.

ST_species.c
* Added an include statement for ST_initialization.h. ST_species.c needs some information about seed dispersal and initialization.

ST_stats.c
* Added an include statement for ST_initialization.h, just like ST_species.c
