+++ ST_globals.h
@@ -29,7 +29,4 @@ extern BmassFlagsType BmassFlags;
 extern MortFlagsType  MortFlags;
 extern GlobalType     SuperGlobals;
 
-extern Bool UseSeedDispersal;
-extern int InitializationMethod;
 extern Bool UseGrid;
-extern Bool DuringInitialization;
